Everyone gets a WorkerService.
For diagnostics I'm forcing the use of WorkerServer regardless of the
users worker coro/class.  You can still bring your own subclass of
WorkerService but if you just use a simple coro it will be wrapped in a
service class.  This changes the signature for coro workers to get a
single argument of WorkerService instance type.

Wiring in some diagnostics and RPC registration from the worker process.
I'm not super stoaked about the way this fans out using lots of little
IPC files but it works for now.